Move some FoundDefinitionRef methods to cc file
